+++ docker/config.py
@@ -30,7 +30,7 @@ DATABASES = {
         'USER': os.getenv('POSTGRES_USER'),
         'PASSWORD': os.getenv('POSTGRES_PASSWORD'),
         'HOST': os.getenv('POSTGRES_HOST'),
-        'PORT': '5432',
+        'PORT': os.getenv('POSTGRES_PORT','5432'),
     }
 }